size

+++ 0x00-hello_world/6-size.c
@@ -1,13 +1,13 @@
 #include <stdio.h>
 /**
  * main - entry point
- * Return: 0 (Succes)
+ * Return: Always 0(Succes)
  */
 int main(void)
 {
-	printf("Size of char: %d byte(s)\n",sizeof(char));
-        printf("Size of int: %d byte(s)\n",sizeof(int));
-	printf("Size of char: %d byte(s)\n",sizeof(long int));
-	printf("Size of char: %d byte(s)\n",sizeof(long long int));
-	printf("Size of char: %d byte(s)\n",sizeof(float));
+	printf("Size of char: %d byte(s)\n", (unsigned long)sizeof(char));
+        printf("Size of int: %d byte(s)\n", (unsigned long)sizeof(int));
+	printf("Size of char: %d byte(s)\n", (unsigned long)sizeof(long int));
+	printf("Size of char: %d byte(s)\n", (unsigned long)sizeof(long long int));
+	printf("Size of char: %d byte(s)\n", (unsigned long)sizeof(float));
 }
